+++ door/__init__.py
@@ -2,8 +2,10 @@ __all__ = (
     'AcquirableDoor',
     'AsyncAcquirableDoor',
     'AsyncSAcquirableDoor',
+    'AsyncSWaitableDoor',
     'AsyncWaitableDoor',
     'SAcquirableDoor',
+    'SWaitableDoor',
     'WaitableDoor',
 )
 
@@ -11,7 +13,9 @@ from door.doors import (
     AcquirableDoor,
     AsyncAcquirableDoor,
     AsyncSAcquirableDoor,
+    AsyncSWaitableDoor,
     AsyncWaitableDoor,
     SAcquirableDoor,
+    SWaitableDoor,
     WaitableDoor,
 )