+++ door/multiprocessing2.py
@@ -1,11 +1,13 @@
 """:mod:`door.multiprocessing2` defines utilities for multiprocessing."""
 
 from dataclasses import dataclass, field
-from multiprocessing import Condition, RLock
+from multiprocessing import Condition, Lock
 
 from door.primitives import (
     Acquirable,
     RSLock as SyncRSLock,
+    SAcquirable,
+    SCondition as SyncSCondition,
     Waitable,
     WSLock as SyncWSLock,
 )
@@ -22,8 +24,8 @@ class RSLock(SyncRSLock):
     and Foundations by Michel Raynal.
     """
 
-    _r: Acquirable = field(default_factory=RLock, init=False)
-    _g: Acquirable = field(default_factory=RLock, init=False)
+    _r: Acquirable = field(default_factory=Lock)
+    _g: Acquirable = field(default_factory=Lock)
 
 
 @dataclass
@@ -33,4 +35,26 @@ class WSLock(SyncWSLock):
     This class is designed to be used for multiprocessing.
     """
 
-    _g: Waitable = field(default_factory=Condition, init=False)
+    _g: Waitable = field(default_factory=Condition)
+
+
+@dataclass
+class RSCondition(SyncSCondition):
+    """The class for read-preferring shared condition variables.
+
+    This class is designed to be used for multiprocessing.
+    """
+
+    _s: SAcquirable = field(default_factory=RSLock)
+    _a: Waitable = field(default_factory=Condition)
+
+
+@dataclass
+class WSCondition(SyncSCondition):
+    """The class for write-preferring shared condition variables.
+
+    This class is designed to be used for multiprocessing.
+    """
+
+    _s: SAcquirable = field(default_factory=WSLock)
+    _a: Waitable = field(default_factory=Condition)